Feat: 채팅방 생성 UI 반응형 구현
