+++ src/pages/HomePage.jsx
@@ -3,13 +3,23 @@ import SideBar from "../components/Main/SideBar";
 import Categories from "../components/Main/Categories";
 import { Spinner } from "@material-tailwind/react";
 import MainPageMain from "../components/Main/MainPageMain";
-import { useRecoilValue } from "recoil";
+import { useRecoilState, useRecoilValue } from "recoil";
 import filterState from "../recoil/atoms/filter";
+import { BsFillChatDotsFill } from "react-icons/bs";
+import createNewChatState from "../recoil/atoms/createNewChat";
+import Button from "../components/Common/Button";
+import useModal from "../hooks/useModal";
+import CreateChatButtonGroup from "../components/Main/CreateChatButtonGroup";
 
 export default function HomePage() {
-  const filter = useRecoilValue(filterState);
+  const [isOpen, handleModalClick] = useModal();
   const [filters, setFilters] = useState([]);
+  const filter = useRecoilValue(filterState);
+  const [createNewChat, setCreateNewChat] = useRecoilState(createNewChatState);
 
+  const handleNewChatClick = () => {
+    setCreateNewChat((prev) => !prev);
+  };
   const handleFilterClick = (value) => {
     if (filters.includes(value)) {
       setFilters((prev) => prev.filter((f) => f !== value));
@@ -24,8 +34,8 @@ export default function HomePage() {
   }, [filter]);
   return (
     <main className="relative flex w-full h-full">
-      <SideBar />
-      <div className="flex flex-col p-4 md:ml-64 w-full">
+      <SideBar isModalOpen={isOpen} onModalClick={handleModalClick} />
+      <div className="relative flex flex-col mb-12 p-4 md:ml-64 w-full">
         <Categories
           filter={filter}
           filters={filters}
@@ -34,7 +44,22 @@ export default function HomePage() {
         <Suspense fallback={<Spinner />}>
           <MainPageMain filters={filters} />
         </Suspense>
+        {createNewChat && (
+          <CreateChatButtonGroup
+            onModalClick={handleModalClick}
+            mobile={false}
+          />
+        )}
       </div>
+      {!createNewChat && (
+        <Button
+          variant="text"
+          className="fixed bottom-8 right-7 md:left-7 md:right-auto p-3 text-white bg-brand hover:bg-white hover:text-brand rounded-full"
+          onClick={handleNewChatClick}
+        >
+          <BsFillChatDotsFill size={24} />
+        </Button>
+      )}
     </main>
   );
 }
