+++ tailwind.config.js
@@ -19,6 +19,7 @@ module.exports = withMT({
       colors: {
         light_black: "rgb(16, 16, 16)",
         brand: "#A66EFC",
+        dark_brand: "#7B5AAC",
         line: "#1F2937",
         dark_gray: "#15171A",
         base_gray: "#424242",
